% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naismithwalkspeed120619.R
\name{naismithwalkspeed}
\alias{naismithwalkspeed}
\title{The speed a person walks based on Naismith's rule}
\usage{
naismithwalkspeed(vo2max, slope)
}
\arguments{
\item{vo2max}{vector, the Vo2max of each individual (ml/kg/min-1)}

\item{slope}{vector, slope of the route as a percentage}
}
\value{
walking speed km/hr
}
\description{
This is a function which estimates how far
a person could walk in one hour (in km/hr)
given the inputs of the person's VO2max and the slope of the hill
It returns the speed
}
\details{
long form explanation of the function
}
\examples{
\dontrun{
naismithwalkspeed(40,4)
naismithwalkspeed(c(40,30),c(4,2))
}
}
