% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naismithwalkspeed120619.R
\name{nplot}
\alias{nplot}
\title{A very simple histogram plot of walk speed}
\usage{
nplot(walk_speed_kmh)
}
\arguments{
\item{walk_speed_kmh}{vector, walking speed km/hr}
}
\value{
a ggplot2 histogram
}
\description{
This is a function which plots a histogram of
the results of naismithwalkspeed using ggplot
}
\details{
long form explanation of the function
}
\examples{
\dontrun{
nplot(c(2,2,4,6,6))
}
}
