% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedal_power250619.R
\name{pedal_power}
\alias{pedal_power}
\title{An estimation of pedalling power}
\usage{
pedal_power(data = NULL, personID = "personID", weight = "weight",
  outputvo2 = "outputvo2", LTpercent = "LTpercent",
  return_a_vector = TRUE)
}
\arguments{
\item{data}{dataframe optional, default is NULL}

\item{personID}{default = "personID" vector or character string of column name in data}

\item{weight}{default = "weight", weight of bike and rider in kg}

\item{outputvo2}{default = "outputvo2", estimate of volume oxygen uptake per min (mL/min)}

\item{LTpercent}{default = "LTpercent", Threshold percentage of Vo2max}

\item{return_a_vector}{default = TRUE}
}
\value{
Threshold percentage of Vo2max
}
\description{
This function allows you to estimate the pedalling power
which a non regular cyclist would be physically capable of for
an hour at a time on a daily basis
It returns pedalling power in Watts
}
\details{
The rationale for this is described in
Philips I, Watling D, Timms P (2018)
Estimating Individual Physical Capability (IPC)
to make journeys by bicycle
International Journal of Sustainable Transportation
http://dx.doi.org/10.1080/15568318.2017.1368748
}
\examples{
\dontrun{
pedal_power(personID = 1,
           weight = 70,
           outputvo2 = 3850,
           LTpercent = 72.5,
           return_a_vector = TRUE)
}
}
