% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vo2max__function250619.R
\name{vo2max}
\alias{vo2max}
\title{An estimate of a person's VO2max}
\usage{
vo2max(data = NULL, personID = "personID", age = "age",
  sex = "sex", weight = "weight", BMI = "BMI", PASS = "PASScat",
  clones = 10, return_a_vector = TRUE)
}
\arguments{
\item{data}{dataframe optional, default is NULL}

\item{personID}{default = "personID" vector or character string of column name in data}

\item{age}{default =  "age",vector or character string of column name in data}

\item{sex}{default = "sex" 0 is male 1 is female}

\item{weight}{default = "weight", weight of bike and rider in kg}

\item{BMI}{default = "BMI", Body Mass Index kg/m2}

\item{PASS}{default = "PASScat", NASA - Physical Activity Survey}

\item{clones}{default  = 10}

\item{return_a_vector}{default = TRUE}
}
\value{
vector default or data frame of estimate of vo2max
}
\description{
This is a function which estimates a persons VO2max
measured in milliliters of oxygen used in one minute per kilogram of body weight (mL/kg/min).
given the inputs of the person's age sex weight BMI and physical activity
It returns measures of VO2max
}
\details{
The function makes estimates of VO2max based on the regression model by Wier et al.
DOI:  10.1249/01.mss.0000193561.64152  given the input values.  This study
aims to be applicable to the population as a whole rather than just a particular
population segment.
#'
}
\examples{
\dontrun{
personID <- c(1,2,3)
age <- c(30,40,50)
sex <- c(0,1,0)
BMI <- c(25,27,30)
weight <- c(50,70,90)
PASScat <- c(7,5,1)
df <- data.frame(personID,age,sex,weight,BMI,PASScat)

test <- vo2max(data= df,personID = "personID",age = "age", sex = "sex", weight = "weight",
              BMI = "BMI",PASS = "PASScat",clones = 1000,return_a_vector = F)
test2 <- vo2max(personID = 1,age = 30, sex = 0, weight = 50,
               BMI = 25,PASS = 7,clones = 10,return_a_vector = F)
test3 <- vo2max(personID = personID,age = age, sex = sex, weight = weight,
               BMI = BMI,PASS = PASScat,clones = 1000,return_a_vector = F)
summary(test$vo2max_norm)
summary(test3$vo2max_norm)
}
}
