% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vo2ThreshPP250619.R
\name{vo2threshpp}
\alias{vo2threshpp}
\title{An estimation of vo2max threshold and output of pedalling power}
\usage{
vo2threshpp(data = NULL, personID = "personID", age = "age",
  sex = "sex", weight = "weight", BMI = "BMI", PASS = "PASScat",
  clones = 10, return_a_vector = TRUE)
}
\arguments{
\item{data}{dataframe optional, default is NULL}

\item{personID}{default = "personID" vector or character string of column name in data}

\item{age}{default =  "age",vector or character string of column name in data}

\item{sex}{default = "sex" 0 is male 1 is female}

\item{weight}{default = "weight", weight of bike and rider in kg}

\item{BMI}{default = "BMI", Body Mass Index kg/m2}

\item{PASS}{default = "PASScat", NASA - Physical Activity Survey}

\item{clones}{default  = 10}

\item{return_a_vector}{default = TRUE}
}
\value{
Vo2max Threshold percentage of Vo2max and pedal pedal power
}
\description{
This function allows you to estimate in one function
the vo2max the Threshold and pedalling power
which a non regular cyclist would be physically capable of for
an hour at a time on a daily basis
It returns Vo2max (mL/kg/min) Threshold (percent)pedalling power(Watts)
}
\details{
The rationale for this is described in
Philips I, Watling D, Timms P (2018)
Estimating Individual Physical Capability (IPC)
to make journeys by bicycle
International Journal of Sustainable Transportation
http://dx.doi.org/10.1080/15568318.2017.1368748
}
\examples{
\dontrun{
vo2threspp(personID = 1,age = 30, sex = 0, weight = 50,
          BMI = 25,PASS = 7,clones = 1,return_a_vector = F)
}
}
